fix(memory): decode INSERT...RETURNING version as i32 not i64
save_compression_guidelines and save_compression_guidelines_with_category
both decoded the INTEGER version column returned by RETURNING directly
into i64 via query_scalar, throwing ColumnDecode on Postgres. Caught by
CI's live-Postgres run of this PR's own new regression test, since Docker
was unavailable during local development.

Same defect class as the rest of this PR; fixed the same way.

+++ crates/zeph-memory/src/store/compression_guidelines.rs
@@ -182,7 +182,10 @@ impl SqliteStore {
         // The INSERT...SELECT computes MAX(version)+1 across all rows (global + per-conversation)
         // and inserts it in a single statement. SQLite's single-writer WAL guarantee makes this
         // atomic — no concurrent writer can observe the same MAX and produce a duplicate version.
-        let new_version: i64 = zeph_db::query_scalar(
+        // `version` is `INTEGER` (`INT4`) on Postgres, so `RETURNING version` decodes as `i32`,
+        // not `i64`; widened back to `i64` below to keep this function's public return type
+        // unchanged.
+        let new_version: i32 = zeph_db::query_scalar(
             sql!("INSERT INTO compression_guidelines (version, guidelines, token_count, conversation_id) \
              SELECT COALESCE(MAX(version), 0) + 1, ?, ?, ? \
              FROM compression_guidelines \
@@ -193,7 +196,7 @@ impl SqliteStore {
         .bind(conversation_id.map(|c| c.0))
         .fetch_one(&self.pool)
         .await?;
-        Ok(new_version)
+        Ok(i64::from(new_version))
     }
 
     /// Log a compression failure pair.
@@ -386,7 +389,10 @@ impl SqliteStore {
         category: &str,
         conversation_id: Option<ConversationId>,
     ) -> Result<i64, MemoryError> {
-        let new_version: i64 = zeph_db::query_scalar(sql!(
+        // `version` is `INTEGER` (`INT4`) on Postgres, so `RETURNING version` decodes as `i32`,
+        // not `i64`; widened back to `i64` below to keep this function's public return type
+        // unchanged.
+        let new_version: i32 = zeph_db::query_scalar(sql!(
             "INSERT INTO compression_guidelines \
              (version, category, guidelines, token_count, conversation_id) \
              SELECT COALESCE(MAX(version), 0) + 1, ?, ?, ?, ? \
@@ -399,7 +405,7 @@ impl SqliteStore {
         .bind(conversation_id.map(|c| c.0))
         .fetch_one(&self.pool)
         .await?;
-        Ok(new_version)
+        Ok(i64::from(new_version))
     }
 
     /// Mark failure pairs as consumed by the updater.
